pyproject.toml
[tool.poetry]
name = "mem0ai"
version = "0.1.49"
description = "Long-term memory for AI Agents"
authors = ["Mem0 <founders@mem0.ai>"]
exclude = [
    "db",
    "configs",
    "notebooks",
    "embedchain",
]
packages = [
    { include = "mem0" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
qdrant-client = "^1.9.1"
pydantic = "^2.7.3"
openai = "^1.33.0"
posthog = "^3.5.0"
pytz = "^2024.1"
sqlalchemy = "^2.0.31"
langchain-community = "^0.3.1"
neo4j = "^5.23.1"
rank-bm25 = "^0.2.2"
python-dotenv = "^1.0.1"

[tool.poetry.extras]
graph = ["langchain-community", "neo4j", "rank-bm25"]

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
isort = "^5.13.2"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = ["embedchain/"]
